requirements.txt
pandas
matplotlib
ta
streamlit
numpy
python-binance
pandas_ta
setuptools
python-dotenv